backend/notifications/requirements.txt
# ==============================================================================
# backend/notifications/requirements.txt
# ==============================================================================

fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

google-cloud-logging==3.11.3
google-cloud-secret-manager==2.21.0

# Email delivery
sendgrid==6.11.0
